Menambahkan API dan Database saldo user

+++ routes/requestMethod.py
@@ -2,14 +2,14 @@
 from fastapi import APIRouter, UploadFile, File
 import secrets
 from PIL import Image
-from models.tabel import users, artikels, tiket, pembelian
+from models.tabel import users, artikels, tiket, pembelian, saldo_user
 from config.db import conn
 from schema.schemas import User, LoginData, CheckCurrentPWD, UpdatePWD, ProfilePicture, BannerPicture, Artikel, UpdateThumbnailArtikel, postArtikel, tiketOrder, beliTiket
 
 
 router = APIRouter()
 
-#! USER
+#! USER  ==============================================================================
 @router.get('/')
 async def fetchAllUserData():
     return conn.execute(users.select()).fetchall()
@@ -57,69 +57,39 @@ async def login(data: LoginData):
 
 @router.post('/register/')
 async def register(user: User):
-    conn.execute(users.insert().values(
-        fullname = user.fullname,
-        email  = user.email,
-        alamat = user.alamat,
-        noTelepon = user.noTelepon,
-        password = user.password,
-        jenisKelamin = user.jenisKelamin
-    ))
+    conn.execute(users.insert().values(fullname = user.fullname,email  = user.email,alamat = user.alamat,noTelepon = user.noTelepon,password = user.password,jenisKelamin = user.jenisKelamin))
     return conn.execute(users.select()).fetchall()
 
 
 @router.put('/update/{id}')
 async def updateUserData(id: int, user: User):
-    conn.execute(users.update().values(
-        fullname = user.fullname,
-        email  = user.email,
-        alamat = user.alamat,
-        noTelepon = user.noTelepon,
-        password = user.password,
-        jenisKelamin = user.jenisKelamin
-    ).where(
-        users.c.id == id
-    ))
+    conn.execute(users.update().values(fullname = user.fullname,email  = user.email,alamat = user.alamat,noTelepon = user.noTelepon,password = user.password,jenisKelamin = user.jenisKelamin).where(users.c.id == id))
     return conn.execute(users.select()).fetchall()
 
 @router.post('/checkcurrentpassword/')
 async def checkCurrentPassword(pwdByID:CheckCurrentPWD):
-    response = conn.execute(users.select().where(
-        users.c.id == pwdByID.id, users.c.password == pwdByID.password 
-    ))
+    response = conn.execute(users.select().where(users.c.id == pwdByID.id, users.c.password == pwdByID.password ))
     if response:
         return response.first()
 
 
 @router.put('/changepassword/{id}')
 async def updatePassword(id: int, pwd: UpdatePWD):
-    response = conn.execute(users.update().values(
-        password =  pwd.password
-    ).where(
-        users.c.id == id
-    ))
+    response = conn.execute(users.update().values(password =  pwd.password).where(users.c.id == id))
     if(response):
         return conn.execute(users.select().where(users.c.id == id)).first()
 
 # *Update profile picture
 @router.put('/profile_picture/{id}')
 async def profilePicture(id: int, picture: ProfilePicture):
-    response = conn.execute(users.update().values(
-        profilPhoto = picture.profilPhoto
-    ).where(
-        users.c.id == id
-    ))
+    response = conn.execute(users.update().values(profilPhoto = picture.profilPhoto).where(users.c.id == id))
     if(response):
         return conn.execute(users.select().where(users.c.id == id)).first()
 
 # *Update profile picture
 @router.put('/banner_picture/{id}')
 async def bannerPicture(id: int, picture: BannerPicture):
-    response = conn.execute(users.update().values(
-        profilBannerPhoto = picture.profilBannerPhoto
-    ).where(
-        users.c.id == id
-    ))
+    response = conn.execute(users.update().values(profilBannerPhoto = picture.profilBannerPhoto).where(users.c.id == id))
     if(response):
         return conn.execute(users.select().where(users.c.id == id)).first()
 
@@ -128,7 +98,7 @@ async def deleteUserData(id: int):
     conn.execute(users.delete().where(users.c.id == id))
     return conn.execute(users.select()).fetchall()
 
-# !ARTIKEL
+# !ARTIKEL  ==============================================================================
 @router.get('/artikelall/')
 async def fetchAllArtikel():
     return conn.execute(artikels.select()).fetchall()
@@ -137,54 +107,33 @@ async def fetchAllArtikel():
 async def artikelOfUser(idOfUser: int):
     return conn.execute(artikels.select().where(artikels.c.user_id == idOfUser)).fetchall()
 
-
 @router.put('/update_artikel/{id}/{idOfUser}')
 async def updateArtikel(id:int, idOfUser: int, update: Artikel):
-    response = conn.execute(artikels.update().values(
-        title = update.title,
-        isi  = update.isi,
-    ).where(
-        artikels.c.id == id, artikels.c.user_id == idOfUser
-    ))
+    response = conn.execute(artikels.update().values(title = update.title, isi  = update.isi,).where(artikels.c.id == id, artikels.c.user_id == idOfUser))
     if(response):
         return conn.execute(artikels.select().where(artikels.c.id == id)).first()
 
 @router.put('/update_thumbnail_artikel/{id}/{idOfUser}')
 async def updateThumbnail(id: int, idOfUser: int, image: UpdateThumbnailArtikel):
-    response = conn.execute(artikels.update().values(
-        thumbnail  = image.thumbnail,
-    ).where(
-        artikels.c.id == id, artikels.c.user_id == idOfUser
-    ))
+    response = conn.execute(artikels.update().values(thumbnail = image.thumbnail,).where(artikels.c.id == id, artikels.c.user_id == idOfUser))
     if(response):
         return conn.execute(artikels.select().where(artikels.c.id == id)).first()
 
 @router.post('/post_artikel/{idOfUser}')
 async def uploadArtikel(idOfUser: int, data: postArtikel):
-    response = conn.execute(artikels.insert().values(
-        user_id = idOfUser,
-        title= data.title,
-        isi= data.isi,
-    ))
+    response = conn.execute(artikels.insert().values( user_id = idOfUser, title= data.title, isi= data.isi,))
     if response:
         return conn.execute(artikels.select().order_by(artikels.c.id.desc())).first()
 
 @router.delete('/delete_artikel/{id}/{idOfUser}')
 async def deleteArtikel(id: int, idOfUser: int):
-    conn.execute(artikels.delete().where(
-        artikels.c.id == id, artikels.c.user_id == idOfUser
-        ))
-    return conn.execute(artikels.select().where(
-        artikels.c.user_id == idOfUser
-        )).fetchall()
+    conn.execute(artikels.delete().where(artikels.c.id == id, artikels.c.user_id == idOfUser))
+    return conn.execute(artikels.select().where(artikels.c.user_id == idOfUser)).fetchall()
 
-# tiket
+# !tiket  ==============================================================================
 @router.post('/order_tiket/')
 async def orderTiket(tiketOrder: tiketOrder):
-    response = conn.execute(tiket.select().where(
-            tiket.c.stasiun_asal == tiketOrder.stasiun_asal, 
-            tiket.c.stasiun_tujuan == tiketOrder.stasiun_tujuan
-            )).first()
+    response = conn.execute(tiket.select().where(tiket.c.stasiun_asal == tiketOrder.stasiun_asal, tiket.c.stasiun_tujuan == tiketOrder.stasiun_tujuan)).first()
     return response
 
 @router.get('/bought_tiket/{tiketID}')
@@ -196,17 +145,28 @@ async def boughtTiket(tiketID: int):
 async def showTiket():
     return conn.execute(tiket.select()).fetchall()
 
-# pembelian
+# !pembelian  ==============================================================================
 @router.post('/pembelian_tiket/')
 async def buyTiket(dataPembeli: beliTiket):
-    response = conn.execute(pembelian.insert().values(
-            user_id = dataPembeli.user_id,
-            tiket_id = dataPembeli.tiket_id,
-            tanggal= dataPembeli.tanggal
-            ))
+    response = conn.execute(pembelian.insert().values( user_id = dataPembeli.user_id, tiket_id = dataPembeli.tiket_id, tanggal= dataPembeli.tanggal ))
     return conn.execute(pembelian.select().where(pembelian.c.user_id == dataPembeli.user_id)).fetchall()
 
 @router.get('/tiket_saya/{userID}')
 async def tiketSaya(userID: int):
     response = conn.execute(pembelian.select().where(pembelian.c.user_id == userID)).fetchall()
     return response
+
+# !Saldo user ==============================================================================
+@router.get('/saldo/')
+async def chekUserSaldo(userID: int):
+    return conn.execute(saldo_user.select().where(saldo_user.c.user_id == userID)).first()
+
+@router.post('/topup/')
+async def  topup( userID: int, nominalSaldo: int):
+    checkAvailabilityUser = conn.execute(saldo_user.select().where(saldo_user.c.user_id == userID)).fetchall()
+    if not checkAvailabilityUser :
+        conn.execute(saldo_user.insert().values(user_id = userID, saldo = nominalSaldo))
+        return conn.execute(saldo_user.select().where(saldo_user.c.user_id == userID)).fetchall()
+    else:
+        conn.execute(saldo_user.update().values(user_id = userID, saldo = nominalSaldo).where(saldo_user.c.user_id == userID))
+        return conn.execute(saldo_user.select().where(saldo_user.c.user_id == userID)).fetchall()
